refactor: 💡 dirs
✅ Closes: #248
