+++ src/features/dump/utils/validate-category.ts
