+++ src/features/submit/components/submit-drawer.tsx
@@ -5,11 +5,11 @@ import {
 	DrawerHeader,
 	DrawerTitle,
 } from "@/components/ui/drawer";
-import { changeBlogStatus } from "@/features/blog/actions/change-blog-status";
+import { changeBlogStatus } from "@/features/dump/actions/change-blog-status";
 import { useToast } from "@/hooks/use-toast";
 import { Link } from "next-view-transitions";
 import { type Dispatch, type SetStateAction, useState } from "react";
-import { changeMypageStatus } from "../../mypage/actions/change-mypage-status";
+import { changeMypageStatus } from "../../dump/actions/change-mypage-status";
 
 type Props = {
 	setOpen: Dispatch<SetStateAction<boolean>>;
